+++ tablet_qt/tasks/rapid3.h
@@ -1,0 +1,72 @@
+/*
+    Copyright (C) 2012-2020 Rudolf Cardinal (rudolf@pobox.com).
+
+    This file is part of CamCOPS.
+
+    CamCOPS is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    CamCOPS is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with CamCOPS. If not, see <http://www.gnu.org/licenses/>.
+*/
+
+#pragma once
+#include <QPointer>
+#include <QString>
+#include "tasklib/task.h"
+
+class CamcopsApp;
+class OpenableWidget;
+class Questionnaire;
+class TaskFactory;
+
+void initializeRapid3(TaskFactory& factory);
+
+
+class Rapid3 : public Task
+{
+    Q_OBJECT
+public:
+    Rapid3(CamcopsApp& app, DatabaseManager& db,
+           int load_pk = dbconst::NONEXISTENT_PK);
+    // ------------------------------------------------------------------------
+    // Class overrides
+    // ------------------------------------------------------------------------
+    virtual QString shortname() const override;
+    virtual QString longname() const override;
+    virtual QString description() const override;
+    virtual TaskImplementationType implementationType() const override {
+        return TaskImplementationType::UpgradableSkeleton;
+    }
+    // ------------------------------------------------------------------------
+    // Instance overrides
+    // ------------------------------------------------------------------------
+    virtual bool isComplete() const override;
+    virtual QStringList summary() const override;
+    virtual QStringList detail() const override;
+    virtual OpenableWidget* editor(bool read_only = false) override;
+    // ------------------------------------------------------------------------
+    // Task-specific calculations
+    // ------------------------------------------------------------------------
+    QVariant rapid3() const;
+    double functionalStatus() const;
+    double painTolerance() const;
+    double globalEstimate() const;
+public:
+    static const QString RAPID3_TABLENAME;
+protected:
+    QPointer<Questionnaire> m_questionnaire;
+private:
+    QStringList allFieldnames() const;
+    QStringList q1Fieldnames() const;
+    QStringList q1ScoringFieldnames() const;
+
+    QString diseaseSeverity() const;
+};